feat: created ResourceCardBody component inside of ResourceCard, and moved ResourceCard from UI to resources folder
