+++ frontend/src/components/resources/ResourceCard.tsx
@@ -1,12 +1,10 @@
 import { useState } from "react";
-import { Calendar } from "lucide-react";
-import ResourceCardHeader from "./ResourceCardHeader";
+import ResourceCardHeader from "../ui/ResourceCardHeader";
+import ResourceCardBody from "../ui/ResourceCardBody";
 import { IntResource } from "../../types";
 import { useUserContext } from "../../context/UserContext";
 import { canBookmark } from "../../data/permission/tempRolesPremission";
-import GenericModal from "./Modal/GenericModal";
-import ContentTypeBadge from "../resources/ContentTypeBadge";
-import heartIcon from "../../assets/heart.svg";
+import GenericModal from "../ui/Modal/GenericModal";
 
 interface ResourceCardProps {
   resource: IntResource;
@@ -60,39 +58,13 @@ const ResourceCard = ({
         handleBookmarkClick={handleBookmarkClick}
         hasBookmarkPermission={hasBookmarkPermission}
       />
-      {tags && (
-        <div className="flex flex-wrap gap-2">
-          {tags.map((tag, index) => (
-            <div
-              key={index}
-              className="px-4 py-1.5 rounded-full border border-gray-200 text-sm font-medium text-gray-700 bg-white"
-            >
-              {typeof tag === "string" ? tag : tag.name}
-            </div>
-          ))}
-        </div>
-      )}
-
-      <div className="flex items-center gap-4 mt-2 text-gray-500 font-medium">
-        <div className="flex items-center gap-1.5">
-          <ContentTypeBadge type={type} />
-        </div>
-
-        <div className="flex items-center gap-1.5">
-          <img
-            src={heartIcon}
-            alt="heart-icon"
-            className="w-5 h-5 opacity-60"
-          />
-          <span className="text-sm">{like_count}</span>
-        </div>
-
-        <div className="flex items-center gap-1.5 ml-auto text-sm">
-          <Calendar size={18} className="text-gray-400" />
-          {formattedDate}
-        </div>
-      </div>
-
+      <ResourceCardBody
+        tags={tags}
+        type={type}
+        likeCount={like_count}
+        formattedDate={formattedDate}
+      />
+      
       <GenericModal
         isOpen={showModal}
         onClose={() => setShowModal(false)}
